chore: logout on expired token
